pyproject.toml

[project]
name = "seattle-2025"
version = "1.0.0"
description = "Convention app for Seattle Worldcon 2025"
authors = []
dependencies = [
    "django-svcs>=0.3.4",
    "granian>=1.6.3",
    "nomnom-hugoawards~=2025.0.2",
    "social-auth-app-django~=5.4",
    "django-fsm>=2.8.1",
    "django-celery>=3.1.17",
    "cryptography>=44.0.0",
    "pyjwt>=2.10.1",
    "sentry-sdk>=2.19.0",
]
requires-python = "~=3.12.3"
readme = "README.md"
license = {text = "MIT"}

[tool.uv.sources]
# nomnom-hugoawards = { path = "../nomnom", editable = true }
# social-auth-core = { path = "../social-core" }

[tool.ruff.lint.isort]
known-first-party = ["", "seattle_2025_app"]

[tool.uv]
dev-dependencies = [
    "djlint>=1.36.4",
    "factory-boy~=3.3.0",
    "ipdb>=0.13.13",
    "pytest~=8.3.3",
    "pytest-check>=2.4.1",
    "pytest-django~=4.9.0",
    "pytest-xdist>=3.6.1",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
looponfailroots = ["config", "seattle_2025_app"]
django_find_project = false
filterwarnings = [
  # I know looponfailroots is 'deprecated' but ... i'm tired of seeing it
  "ignore::DeprecationWarning:xdist.plugin"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["seattle_2025_app"]
